(#87) Added explicit type cast to calculation of allowed skew amount

+++ tests/c++tests/test-os/src/test-timer.cxx
@@ -120,8 +120,8 @@ namespace
     OS::ElapsedTime e = t.getElapsed();
     unsigned int msecs = e.seconds() * 1000 + e.nanoseconds() / 1000000;
     // Allow for a 3% skew
-    EXPECT_GE( msecs, run_time * 0.97 );
-    EXPECT_LE( msecs, run_time * 1.03 );
+    EXPECT_GE( msecs, (unsigned int)(run_time * 0.97) );
+    EXPECT_LE( msecs, (unsigned int)(run_time * 1.03) );
   }
